Modified GameProfileController. Added list to index view, aded function for the route for creating new game profile. Added function for storing new profile creation. Mofified Profile model. Changed casts field. added route for navigation in navigation2.blade.php. added create blade php. added form fields and route for saving profile data. Modified index.php. Printed values from the list passed by controller. Added pagination. In web. php, added routes for create and store new game profile data.

+++ game-statistics/app/Models/Profile.php
@@ -13,8 +13,8 @@ class Profile extends Model
         "sequel_id"
     ];
     protected $casts = [
-        "game_id"=>'numeric',
-        "sequel_id"=>'numeric'
+        "game_id"=>'integer',
+        "sequel_id"=>'integer'
     ];
     public function game(){
         return $this->belongsTo(Game::class);
